Correct double/float type mixups

See merge request gtd-gmbh/libmcs!121

+++ libm/common/tools.h
@@ -356,7 +356,7 @@ static inline double __raise_inexact(double x) {
 }
 
 static inline float __raise_invalidf() {
-    double r = __forced_calculationf(0.0f / 0.0f);
+    float r = __forced_calculationf(0.0f / 0.0f);
     return r;
 }
 static inline float __raise_div_by_zerof(float x) {
